refactoring : 좀 더 명확한 구조를 위한 메서드 이동
